hwdata: remove restriction to platforms.linux
hwdata is just a collection of (plain-text) databases and hence makes
sense on any platform. The impetus for this commit was #112644, where we
wanted to change pciutils so that it uses the database from hwdata. At
this time, pciutils was marked as platforms.unix but hwdata only as
platforms.linux.